requirements.txt
distribute==0.6.14
wsgiref==0.1.2
Django==1.3
psycopg2==2.4
PyYAML==3.09
git+http://github.com/facebook/python-sdk.git#egg=facebook-python-sdk
South==0.7.3
django-social-auth==0.3.9
httplib2==0.6.0
oauth2==1.5.167
python-openid==2.2.5
Fabric==1.0.1
